[partial] Refactor, add tests. A few new features.
This is a shit commit, but screw it, I'm tired.

Refactor some of the bits out into classes. This is by no means done, but
it will start letting me add default guesses to things the user does not
feel like specifying, so I think it'll be worth it.

Some dogshit tests and some scaffolding to make it easier to test things.

I added a "right axis" feature. I'm not sure I 100% understand how this
works inside rrdtool, but it seems to do the job for me.

I found a couple of bugs, and fixed one (where you use multiple RRD files
which have the same DS inside of them, it would break.. you can now avoid
this with the `index` property of each line, just make sure that's
unique.

It defaults to slope-mode now, because I think that's prettier... it,
along with things like turning off the rrdtool banner, will all be
configurable later.

It supports time zones, by clobbering the environment before calling
rrdtool. IDK if there's a better way, but this works.

I'm pretty sure the config.yml.example is broken, but I'll fix that
later.