Add more tests, and some exceptions for common mistakes.
The entire `draw_one` function needs breaking up, and I'm not convinced
my design is sound at all, but for now let's start testing things and
gracefully catching common mistakes.

We'll check for a few of these mistakes, and we've got some shiny new
library-specific exceptions that we can raise if we find one. The default
script just catches all of these (they inherit from a common ancestor),
but I'd like to be friendly if I decide to use this code in another
package so we won't just print the error messages inside the library.

Still got lots more testing to do - the coverage report lies, it says
97% coverage but the actual testing is bugger all.