Add support for HRULEs.
Each graph now takes a list of items in the "hrules" section, with the
following arguments:

- position: the vertical position of the line.
- color: a html-style hex-triplet color code.
- legend: Text that should go in the legend at the bottom of the graph.

Legend is optional, the other two are required - if you're missing one
the line will not be drawn.

I think this can be done better, but for now it scratched an itch I had,
and I don't think I've painted myself into a corner as far as backwards
compatibility goes.