Support null legends to leave a line off the legend.
This commit lets me set `legend: false` on a line in order to not show
it on the legend. What's the use of this? You can use it for stacking
lines/areas to show sums, without using VDEFs (which I currently have
not thought of a way to support yet)... because if you just naively
stack a line on top of other, the values shown in the legend become
incorrect.

For my usage, I graph the outputs of my two solar PV systems separately,
then draw another set of them as an area, with the second one stacked on
the first, to show the sum of it.

This code is dogshit and it really shows that this whole function needs
refactoring to clean it up, but it works and I'm tired so I'm committing
it.