Fix field widths in legends?
This code is obscene, I'm not even 100% sure it works correctly, christ
what a mess! So the issues I was having with the field widths was due to
variable-sized units, but it was further complicated by percentages in
them - for instance '%S' counts as two characters in the configuration,
but in practice it's *mostly* one character. Things like kibibytes, if
RRDtool even supports them, will still break this but I'm not sure how to
solve that one.

So now what we're doing is I've removed virtually *all* the manual
spacing from the legend, and we calculate the longest of the legend
names, field width (which is no longer hard-coded at 12, I can override
this if I want), and the length of the units. The presence of any '%'
character subtracts one from the unit length.

So then we add all this up, and space things out, subtracting the width
of the column names if required, and from my limited testing it seems
like it "just works", which is highly suspicious.