Fix #1 - don't choke on colons in legend fields
Fixes: #1

The presence of a colon in any of the legends would be interpreted by
RRDtool as a field separator instead of literal text. We can solve this
by escaping any just before it's passed through.

We don't want to do this before the size calculations though, or the
column widths would be off.

So far the only two places I'm aware of where this is a problem are in
lines and hrules.